+++ NEWS.md
@@ -8,6 +8,9 @@
 -   Added option for users to specify custom outcome models in the
     `gformula()` function. See the vignette “Using Custom Outcome Models
     in gfoRmula”.
+-   Added the option to not truncate the covariates simulated from a
+    normal distribution. See the argument `sim_trunc` to the
+    `gformula()` function
 -   Fixed an issue where the point estimates differed when changing the
     number of bootstrap samples. Since this fix involved adding a
     `set.seed` statement, point estimates can be numerically different